server/requirements.txt
accelerate
bitsandbytes
numpy
opencv-python
pillow
python-rtmidi
requests
scamp
scamp_extensions
swarms
torch
transformers